make building image and setting infra optional
