+++ test/utils/e2eutils.go
@@ -41,14 +41,12 @@ import (
 	promAPI "github.com/prometheus/client_golang/api"
 	promv1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	"github.com/prometheus/common/model"
-	appsv1 "k8s.io/api/apps/v1"
 	autoscalingv1 "k8s.io/api/autoscaling/v1"
 	autoscalingv2 "k8s.io/api/autoscaling/v2"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/utils/ptr"
@@ -825,101 +823,6 @@ func LogVariantAutoscalingStatus(ctx context.Context, vaName, namespace string, 
 	return nil
 }
 
-// creates a llm-d-sim deployment with the specified configuration
-func CreateLlmdSimDeployment(namespace, deployName, modelName, appLabel, port string, avgTTFT, avgITL int, replicas int32) *appsv1.Deployment {
-	return &appsv1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      deployName,
-			Namespace: namespace,
-		},
-		Spec: appsv1.DeploymentSpec{
-			Replicas: ptr.To(replicas),
-			Selector: &metav1.LabelSelector{
-				MatchLabels: map[string]string{
-					"app":                       appLabel,
-					"llm-d.ai/inferenceServing": "true",
-					"llm-d.ai/model":            "ms-sim-llm-d-modelservice",
-				},
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{
-						"app":                       appLabel,
-						"llm-d.ai/inferenceServing": "true",
-						"llm-d.ai/model":            "ms-sim-llm-d-modelservice",
-					},
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Name:            appLabel,
-							Image:           "ghcr.io/llm-d/llm-d-inference-sim:latest",
-							ImagePullPolicy: corev1.PullAlways,
-							Args: []string{
-								"--model",
-								modelName,
-								"--port",
-								port,
-								fmt.Sprintf("--time-to-first-token=%d", avgTTFT),
-								fmt.Sprintf("--inter-token-latency=%d", avgITL),
-								"--mode=random",
-							},
-							Env: []corev1.EnvVar{
-								{Name: "POD_NAME", ValueFrom: &corev1.EnvVarSource{
-									FieldRef: &corev1.ObjectFieldSelector{
-										APIVersion: "v1",
-										FieldPath:  "metadata.name",
-									},
-								}},
-								{Name: "POD_NAMESPACE", ValueFrom: &corev1.EnvVarSource{
-									FieldRef: &corev1.ObjectFieldSelector{
-										APIVersion: "v1",
-										FieldPath:  "metadata.namespace",
-									},
-								}},
-							},
-							Ports: []corev1.ContainerPort{
-								{ContainerPort: 8000, Name: appLabel, Protocol: corev1.ProtocolTCP},
-							},
-						},
-					},
-					RestartPolicy: corev1.RestartPolicyAlways,
-				},
-			},
-		},
-	}
-}
-
-// creates a service for the llm-d-sim deployment
-func CreateLlmdSimService(namespace, serviceName, appLabel string, nodePort, port int) *corev1.Service {
-	return &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      serviceName,
-			Namespace: namespace,
-			Labels: map[string]string{
-				"app":                       appLabel,
-				"llm-d.ai/inferenceServing": "true",
-			},
-		},
-		Spec: corev1.ServiceSpec{
-			Selector: map[string]string{
-				"app":                       appLabel,
-				"llm-d.ai/inferenceServing": "true",
-			},
-			Ports: []corev1.ServicePort{
-				{
-					Name:       appLabel,
-					Port:       int32(port),
-					Protocol:   corev1.ProtocolTCP,
-					TargetPort: intstr.FromInt32(int32(port)),
-					NodePort:   int32(nodePort),
-				},
-			},
-			Type: corev1.ServiceTypeNodePort,
-		},
-	}
-}
-
 // creates a VariantAutoscaling resource with owner reference to deployment
 func CreateVariantAutoscalingResource(namespace, resourceName, modelId, acc string, variantCost float64) *v1alpha1.VariantAutoscaling {
 	return &v1alpha1.VariantAutoscaling{
@@ -936,37 +839,6 @@ func CreateVariantAutoscalingResource(namespace, resourceName, modelId, acc stri
 				Name: resourceName, // Use resourceName as the deployment name
 			},
 			ModelID: modelId,
-		},
-	}
-}
-
-// creates a ServiceMonitor for llm-d-sim metrics collection
-func CreateLlmdSimServiceMonitor(name, namespace, targetNamespace, appLabel string) *promoperator.ServiceMonitor {
-	return &promoperator.ServiceMonitor{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-			Labels: map[string]string{
-				"app":     appLabel,
-				"release": "kube-prometheus-stack",
-			},
-		},
-		Spec: promoperator.ServiceMonitorSpec{
-			Selector: metav1.LabelSelector{
-				MatchLabels: map[string]string{
-					"app": appLabel,
-				},
-			},
-			Endpoints: []promoperator.Endpoint{
-				{
-					Port:     appLabel,
-					Path:     "/metrics",
-					Interval: promoperator.Duration("15s"),
-				},
-			},
-			NamespaceSelector: promoperator.NamespaceSelector{
-				MatchNames: []string{targetNamespace},
-			},
 		},
 	}
 }
